+++ Sources/Support.swift
@@ -31,6 +31,8 @@ enum SettingsKeys {
     static let reasoningInline = "reasoningInline"
     static let specMTP = "specMTP"
     static let faAmd = "faAmd"
+    static let persistCache = "persistCache"
+    static let multiGPU = "multiGPU"
     static let apiKeyEnabled = "apiKeyEnabled"
     static let menuBarIcon = "menuBarIcon"
     static let autoStart = "autoStart"